requirements.txt
playwright==1.41.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
python-telegram-bot==20.7
apscheduler==3.10.4
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
python-multipart==0.0.6
httpx~=0.25.2
aiofiles==23.2.1
pillow==10.2.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0
pandas==2.2.0
matplotlib==3.8.2
